pyproject.toml
[tool.poetry]
name = "llm finetuning"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
transformers = "4.34.0"
torch = "^2.1.1"
torchaudio = "^2.1.1"
torchvision = "^0.16.1"
bitsandbytes = "^0.41.2.post2"
peft = "^0.6.2"
accelerate = "^0.25.0"
lightning = "^2.1.2"
pandas = "^2.1.3"
ipykernel = "^6.27.1"
click = "^8.1.7"
matplotlib = "^3.8.2"
kaggle = "^1.5.16"
scipy = "^1.11.4"
pydantic = "^2.5.2"
llama-cpp-python = "^0.2.20"
pyyaml = "^6.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
